pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "tornado-webhook"
description = "Tornado Webhook is a web application to execute webhook actions."
readme = "README.md"
requires-python = ">=3.7"
license = "MIT"
keywords = ["tornado", "webhook"]
authors = [
  { name = "Wentao Han", email = "wentao.han@gmail.com" },
]
classifiers = [
  "Development Status :: 4 - Beta",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3.7",
  "Programming Language :: Python :: 3.8",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: Implementation :: CPython",
  "Programming Language :: Python :: Implementation :: PyPy",
]
dependencies = ["tornado", "pyyaml"]
dynamic = ["version"]

[project.urls]
Documentation = "https://github.com/hanwentao/tornado-webhook#readme"
Issues = "https://github.com/hanwentao/tornado-webhook/issues"
Source = "https://github.com/hanwentao/tornado-webhook"

[tool.hatch.version]
path = "tornado_webhook/__about__.py"

[tool.hatch.build.targets.sdist]
[tool.hatch.build.targets.wheel]

[tool.hatch.envs.default]
dependencies = [
  "pytest",
  "pytest-cov",
]
[tool.hatch.envs.default.scripts]
cov = "pytest --cov-report=term-missing --cov-config=pyproject.toml --cov=tornado_webhook --cov=tests"
no-cov = "cov --no-cov"

[[tool.hatch.envs.test.matrix]]
python = ["37", "38", "39", "310"]

[tool.coverage.run]
branch = true
parallel = true
omit = [
  "tornado_webhook/__about__.py",
]

[tool.coverage.report]
exclude_lines = [
  "no cov",
  "if __name__ == .__main__.:",
  "if TYPE_CHECKING:",
]
